requirements.txt
opencv-python~=4.7.0.72
ultralytics~=8.0.145
Pillow~=9.5.0
roboflow~=1.1.6
openvino~=2023.1.0
numpy~=1.21.6